[lpgbt_sc_worker] implementation of multi-command/reply handling for slow control speed-up

+++ HWInterface/D19cCommandProcessorInterface.cc
@@ -17,7 +17,7 @@ D19cCommandProcessorInterface::~D19cCommandProcessorInterface() {}
 void D19cCommandProcessorInterface::WriteCommand(const std::vector<uint32_t>& pCommand)
 {
     std::lock_guard<std::recursive_mutex> theGuard(fMutex);
-    uint8_t                               cWordIndex = 0;
+    int                               cWordIndex = 0;
     for(auto cWord: pCommand)
     {
         LOG(DEBUG) << GREEN << "\t Write command word " << +cWordIndex << " value 0x" << std::setfill('0') << std::setw(8) << std::hex << +cWord << RESET;
@@ -30,7 +30,7 @@ std::vector<uint32_t> D19cCommandProcessorInterface::ReadReply(uint8_t pNWords)
 {
     std::lock_guard<std::recursive_mutex> theGuard(fMutex);
     std::vector<uint32_t>                 cReply     = ReadBlockReg("fc7_daq_ctrl.command_processor_block.cpb_reply_fifo", pNWords);
-    uint8_t                               cWordIndex = 0;
+    int                               cWordIndex = 0;
     for(auto cWord: cReply)
     {
         LOG(DEBUG) << YELLOW << "\t Read reply word " << +cWordIndex << " value 0x" << std::setfill('0') << std::setw(8) << std::hex << +cWord << RESET;
